move session type to session package
